+++ viburnum/application/__init__.py
@@ -1,14 +1,16 @@
 from .base import Application, Handler, Resource, ResourceConnector
-from .connectors import SqsPermission, sqs
+from .connectors import S3Permission, SqsPermission, s3, sqs
 from .handlers import (
     JobEvent,
     QueueEvent,
     Request,
     Response,
+    S3EventSequence,
     SqsEventsSequence,
     SqsFailedEvents,
     job,
     route,
+    s3_handler,
     sqs_handler,
 )
-from .resources import Sqs
+from .resources import S3, Sqs